Update DissonanceHeights.html
Below is an updated version of your code. The main fixes include:

Defining the missing navigation functions:
I added functions for showPowerPanel(), showGarbagePanel(), and others so that the inline onclick handlers can call them.

Ensuring proper script order:
The Firebase SDKs are loaded in the head so that the global firebase object is available, and then all of your functions (including the navigation and chat functions) are defined in a single script at the end of the body.

No socket code is used:
The chat is handled entirely via Firestore. (If you need a persistent history for each room, you’ll have to add an index in the Firebase Console as prompted.)